+++ src/main/java/open/commons/spring/web/servlet/InternalServerException.java
@@ -26,6 +26,8 @@
 
 package open.commons.spring.web.servlet;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
 
 /**
@@ -35,6 +37,7 @@ import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExcep
  * @version
  * @author Park_Jun_Hong_(parkjunhong77@gmail.com)
  */
+@ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
 public class InternalServerException extends RuntimeException {
     /**
      *